Created content to conform to README best practices template, removed extraneous READMEs, and added markdown extension
Before alterations, this module contained no README.

As part of a joint Forge/Docs team effort to standardize formatting and
encourage quality module documentation, a best practices README
template was created via internal and external user testing. That
template was then applied to this module.

I created content through testing and deploying the module in a test
environment and documenting the steps. Standard headings were added
(Overview, Module Description, Setup, Usage, etc.) to organize content,
and basic formatting was done to adhere to template standards. File
extension .md was added to enable Markdown formatting in GitHub. As
this module was built with the pre-faces version of the module tool,
READMEs explaining each folder's purpose were still present. Deleted
those for consistency and to help the module get with the times.