(MODULES-7069) Add support for AIX (#115)
* (MODULES-7069) Add support for AIX

Prior to this commit this module had a noop for AIX. This commit adds
support for AIX to specify an `/etc/motd`, `/etc/issue` and
`/etc/issue.net`.

* (FM-6043) Manage file owner and group

Prior to this commit, the files did nto specify the owner or group for
the files that it manages. This commit manages the file owner and group
for all files in the module.

* (MAINT) Fix Travis release_check tests

Prior to this commit, the travis release_check tests were failing due to
a deprecation warning. This commit updates the .sync.yml to specify the
puppet 4.x version when using ruby 2.1.9.