Dont install litmus on puppet 5 ruby version 2.5.1
This is not a good idea, however it is one of the last options we have
to progress. There are gem dependencies when running litmus, pdk and
bolt on puppet 5. Doing this means litmus will not be installed on
puppet 5. We have pinned to a ruby version as puppet and ruby are
closely coupled and this is a solution that works and the ruby version
is available to us.