(MAINT) Stale bot message bot_update
Prior to this commit, the stale bot would use a default message when
marking a community PR as 'stale'. This message wasn't very
descriptive and sometimes would come as too aggressive.

This commit aims to fix the stale bot by giving it a clearer
description of its role and job, as well as the situation of the PR
being marked.