(GH-422) Fix broken spec tests
Prior to this commit spec tests were failing because a change was
introduced that used hash map style sytax for retrieving facts.

The existing tests did not cater for this and would cause failures
because the requested fact was not defined.

This commit fixes the issue by adding an OS has in to the defined facts
for the FreeBSD spec.