(CAT-1805) Update tests to account for PDK integration
This module is currently used in oder to test the pdk, with commands being run against it.
As part of this however the values returned by the tests I have updated here can vary between the PDK integration test suite on Jenkins and the MOTD test suite on Github Actions.
This change aims to resolve this by generalizing the expected values that can very between them.